Extract rateinfo into own file.
