+++ waves/src/Trade.tsx
@@ -10,6 +10,7 @@ import calculateBetaAmount, { getDirection } from "./calculateBetaAmount";
 import AssetSelector from "./components/AssetSelector";
 import ExchangeIcon from "./components/ExchangeIcon";
 import { makeBuyCreateSwapPayload, makeSellCreateSwapPayload, signAndSend, WalletStatus } from "./wasmProxy";
+import RateInfo from "./components/RateInfo";
 
 const debug = Debug("Swap");
 const error = Debug("Swap:error");
@@ -205,22 +206,5 @@ function BlockExplorerLink() {
     </Link>;
 }
 
-interface RateInfoProps {
-    rate: Rate;
-    direction: "ask" | "bid";
-}
-
-function RateInfo({ rate, direction }: RateInfoProps) {
-    switch (direction) {
-        case "ask":
-            return <Box>
-                <Text textStyle="smGray">{rate.ask} USDT ~ 1 BTC</Text>
-            </Box>;
-        case "bid":
-            return <Box>
-                <Text textStyle="smGray">1 BTC ~ {rate.bid} USDT</Text>
-            </Box>;
-    }
-}
 
 export default Trade;
